fix: resolve node:fs bundler errors via browser export condition
Bundlers (webpack, Vite, Turbopack) statically analyse dynamic imports
regardless of runtime guards, causing the node:fs import in the WASM
loader to be pulled into browser bundles.

- Split src/web into shared.ts (API factory), index.node.ts (fs-based
  loader), and index.browser.ts (CDN-only, no node:fs)
- Move expo source to src/expo/ for clearer platform separation
- Add `browser` export condition in package.json routing bundlers to
  the fs-free index.browser.mjs build
- Update tsdown config, tsconfigs, and example path aliases accordingly
