+++ tsconfig.lib.json
@@ -12,6 +12,6 @@
     "skipLibCheck": true
   },
   "include": [
-    "src/index.ts"
+    "src/web"
   ]
 }
